selectors live in redux folder, reducers being imported properly, extract logger settings
