+++ src/js/redux/loggerSettings.js
@@ -1,0 +1,6 @@
+module.exports = {
+  collapsed: true,
+  predicate: (getState, action) =>
+    action.type !== 'CHECK_TILE' &&
+    action.type !== 'WEATHER_POWER'
+}
