+++ src/js/redux/root_reducer.js
@@ -1,16 +1,16 @@
 import { combineReducers } from 'redux'
-import text from '../modules/text.js'
-import isDragging from '../modules/setDrag.js'
-import updating from '../modules/isUpdating.js'
-import weather from '../modules/weatherPower.js'
-import score from '../modules/score.js'
-import view from '../modules/view.js'
-import moves from '../modules/moves.js'
-import board from '../modules/board.js'
-import fallingMagnitudeArray from '../modules/fallingMagnitude.js'
-import growingMoves from '../modules/growingMoves.js'
-import isLeavingArray from '../modules/isLeaving.js'
-import isEnteringArray from '../modules/isEntering.js'
+import text from './modules/text.js'
+import isDragging from './modules/setDrag.js'
+import updating from './modules/isUpdating.js'
+import weather from './modules/weatherPower.js'
+import score from './modules/score.js'
+import view from './modules/view.js'
+import moves from './modules/moves.js'
+import board from './modules/board.js'
+import fallingMagnitudeArray from './modules/fallingMagnitude.js'
+import growingMoves from './modules/growingMoves.js'
+import isLeavingArray from './modules/isLeaving.js'
+import isEnteringArray from './modules/isEntering.js'
 
 const rootReducer = combineReducers({
   text,
