+++ src/js/redux/selectors/selector_isDraggingArray.js
@@ -1,6 +1,6 @@
 import { createSelector } from 'reselect'
-import { booleanArray, transformTiles } from '../model/model.js'
-import { falseBoard } from '../model/constants.js'
+import { booleanArray, transformTiles } from '../../model/model.js'
+import { falseBoard } from '../../model/constants.js'
 
 const _moveArray = (state) => state.moves.moveArray
 const _board = (state) => state.board
